move golden and tests to p2p syncer
